requirements.txt
streamlit
pandas
numpy
matplotlib
websocket-client